requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
qrcode==8.0
Pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
